Move tests for remote browser behavior to main Puppeteer test file. Avoiding unnecessary multiple execution of unit/runner tests with Puppeteer helper.
